Settings reorg and minor cleanup
