+++ app/src/main/java/moe/apex/breadboard/preferences/Preferences.kt
@@ -1,90 +1,57 @@
 package moe.apex.breadboard.preferences
 
-import android.net.Uri
-import android.provider.OpenableColumns
-import androidx.compose.animation.AnimatedVisibility
-import androidx.compose.animation.animateContentSize
-import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.background
+import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
+import androidx.core.graphics.ColorUtils
 import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.rounded.Dashboard
 import androidx.compose.material.icons.rounded.Info
 import androidx.compose.material.icons.rounded.MoreVert
 import androidx.compose.material.icons.rounded.Science
-import androidx.compose.material3.AlertDialog
-import androidx.compose.material3.Button
+import androidx.compose.material.icons.rounded.Settings
+import androidx.compose.material.icons.rounded.Storage
+import androidx.compose.material.icons.rounded.Visibility
 import androidx.compose.material3.DropdownMenu
 import androidx.compose.material3.DropdownMenuItem
 import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.ExperimentalMaterial3ExpressiveApi
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
-import androidx.compose.material3.TextButton
 import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.material3.rememberTopAppBarState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberCoroutineScope
-import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
-import androidx.compose.runtime.toMutableStateList
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.input.nestedscroll.nestedScroll
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.text.LinkAnnotation
-import androidx.compose.ui.text.SpanStyle
-import androidx.compose.ui.text.TextLinkStyles
-import androidx.compose.ui.text.buildAnnotatedString
-import androidx.compose.ui.text.input.KeyboardType
-import androidx.compose.ui.text.style.TextDecoration
-import androidx.compose.ui.text.withLink
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavHostController
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.launch
-import kotlinx.coroutines.withContext
-import moe.apex.breadboard.image.ImageBoard
-import moe.apex.breadboard.image.ImageBoardAuth
-import moe.apex.breadboard.image.ImageBoardRequirement
 import moe.apex.breadboard.navigation.AboutSettings
-import moe.apex.breadboard.navigation.BlockedTagsSettings
+import moe.apex.breadboard.navigation.ContentSettings
+import moe.apex.breadboard.navigation.DataSettings
 import moe.apex.breadboard.navigation.ExperimentalSettings
-import moe.apex.breadboard.navigation.RecommendationsSettings
-import moe.apex.breadboard.prefs
-import moe.apex.breadboard.util.AgeVerification
+import moe.apex.breadboard.navigation.GeneralSettings
+import moe.apex.breadboard.navigation.LayoutSettings
 import moe.apex.breadboard.util.ChevronRight
-import moe.apex.breadboard.util.ExportDirectoryHandler
-import moe.apex.breadboard.util.ExpressiveGroup
+import moe.apex.breadboard.util.LazyExpressiveGroup
 import moe.apex.breadboard.util.MainScreenScaffold
-import moe.apex.breadboard.util.StorageLocationSelection
-import moe.apex.breadboard.util.ImportException
-import moe.apex.breadboard.util.ImportHandler
-import moe.apex.breadboard.util.LARGE_SPACER
-import moe.apex.breadboard.util.PromptType
 import moe.apex.breadboard.util.MEDIUM_SPACER
-import moe.apex.breadboard.util.SMALL_SPACER
-import moe.apex.breadboard.util.TINY_SPACER
 import moe.apex.breadboard.util.TitleSummary
-import moe.apex.breadboard.util.exportData
-import moe.apex.breadboard.util.importData
-import moe.apex.breadboard.util.launchInWebBrowser
-import moe.apex.breadboard.util.preImportChecks
-import moe.apex.breadboard.util.saveUriToPref
-import moe.apex.breadboard.util.showToast
-import moe.apex.breadboard.viewmodel.getGlobalViewModel
-import org.json.JSONException
-import org.json.JSONObject
-import java.io.FileInputStream
 
 
 @OptIn(ExperimentalMaterial3Api::class)
@@ -92,141 +59,6 @@ import java.io.FileInputStream
 fun PreferencesScreen(navController: NavHostController) {
     val topAppBarState = rememberTopAppBarState()
     val scrollBehavior = TopAppBarDefaults.exitUntilCollapsedScrollBehavior(topAppBarState)
-    val viewModel = getGlobalViewModel()
-    val context = LocalContext.current
-    val scope = rememberCoroutineScope()
-
-    var storageLocationPromptLaunched by remember { mutableStateOf(false) }
-    var showExportDialog by remember { mutableStateOf(false) }
-    var exportedData: JSONObject? by remember { mutableStateOf(null) }
-    var importedData: JSONObject? by remember { mutableStateOf(null) }
-    var importingStarted by rememberSaveable { mutableStateOf(false) }
-    var showAuthDialog by remember { mutableStateOf(false) }
-    var showAgeVerificationDialog by remember { mutableStateOf(false) }
-
-    val preferencesRepository = LocalContext.current.prefs
-    val currentSettings = LocalPreferences.current
-
-    if (showAuthDialog) {
-        AuthDialog(
-            selectedBoard = currentSettings.imageSource.imageBoard,
-            default = currentSettings.authFor(currentSettings.imageSource, context),
-            onDismissRequest = { showAuthDialog = false }
-        ) { username, apiKey ->
-            scope.launch {
-                preferencesRepository.setAuth(
-                    currentSettings.imageSource,
-                    username.takeUnless { it.isBlank() },
-                    apiKey.takeUnless { it.isBlank() }
-                )
-            }
-            showAuthDialog = false
-            viewModel.resetProviders()
-        }
-    }
-
-    if (showAgeVerificationDialog) {
-        AgeVerification.AgeVerifyDialog(
-            onDismissRequest = { showAgeVerificationDialog = false },
-            onAgeVerified = { showAgeVerificationDialog = false }
-        )
-    }
-
-    if (storageLocationPromptLaunched) {
-        StorageLocationSelection(
-            promptType = PromptType.DIRECTORY_PERMISSION,
-            onFailure = { storageLocationPromptLaunched = false }
-        ) { uri ->
-            saveUriToPref(context, scope, uri)
-            storageLocationPromptLaunched = false
-        }
-    }
-
-    if (showExportDialog) {
-        val categories = remember { PrefCategory.entries.toMutableStateList() }
-        ExportDialog(
-            enabledCategories = categories,
-            onDismissRequest = { showExportDialog = false }
-        ) {
-            scope.launch {
-                exportedData = exportData(context, categories)
-                showExportDialog = false
-            }
-        }
-    }
-
-    if (exportedData != null) {
-        ExportDirectoryHandler(exportedData!!) { exportedData = null }
-    }
-
-    if (importingStarted) {
-        ImportHandler({ importingStarted = false }) { uri ->
-            var displayName = uri.lastPathSegment.toString()
-            if (!displayName.endsWith(".bread")) {
-                /* Android's content provider also sometimes returns URIs that hide the real file
-                   path and name so we need to query the content resolver to get the correct one. */
-                context.contentResolver.query(uri, null, null, null).use { cursor ->
-                    if (cursor?.moveToFirst() == true) {
-                        val displayNameIndex = cursor.getColumnIndex(OpenableColumns.DISPLAY_NAME)
-                        if (displayNameIndex != -1) {
-                            displayName = cursor.getString(displayNameIndex)
-                        }
-                    }
-                }
-            }
-            if (!displayName.endsWith(".bread")) {
-                showToast(context, "Invalid file type.")
-                importingStarted = false
-                return@ImportHandler
-            }
-
-            scope.launch(Dispatchers.IO) {
-                context.contentResolver.openFileDescriptor(uri, "r").use { fd ->
-                    FileInputStream(fd!!.fileDescriptor).use { fis ->
-                        try {
-                            val json = JSONObject(fis.readBytes().decodeToString())
-                            val checkResult = preImportChecks(currentSettings, json)
-                            if (!checkResult.isSuccess) {
-                                throw checkResult.exceptionOrNull()!!
-                            }
-                            importedData = json
-                        } catch (e: Exception) {
-                            withContext(Dispatchers.Main) {
-                                showToast(
-                                    context = context,
-                                    text = when (e) {
-                                        is JSONException -> "Invalid backup provided."
-                                        is ImportException -> e.message!!
-                                        else -> "Unknown error."
-                                    }
-                                )
-                            }
-                        }
-                    }
-                }
-                importingStarted = false
-            }
-        }
-    }
-
-    if (importedData != null) {
-        ImportDialog(
-            allowedCategories = PrefCategory.entries.filter { pc -> pc == PrefCategory.BUILD || pc.name in importedData!! },
-            onDismissRequest = { importedData = null}
-        ) { categories, merge ->
-            scope.launch {
-                val result = importData(context, importedData!!, categories, merge)
-                withContext(Dispatchers.Main) {
-                    if (result.isFailure) {
-                        showToast(context, result.exceptionOrNull()!!.message!!)
-                    } else {
-                        showToast(context, "Imported successfully.")
-                    }
-                }
-                importedData = null
-            }
-        }
-    }
 
     MainScreenScaffold(
         title = "Settings",
@@ -243,22 +75,9 @@ fun PreferencesScreen(navController: NavHostController) {
                 DropdownMenu(
                     expanded = isDropdownVisible,
                     containerColor = MaterialTheme.colorScheme.surfaceContainerHigh,
-                    shape = MaterialTheme.shapes.small,
+                    shape = MaterialTheme.shapes.medium,
                     onDismissRequest = { isDropdownVisible = false }
                 ) {
-                    DropdownMenuItem(
-                        text = { Text("About") },
-                        leadingIcon = {
-                            Icon(
-                                imageVector = Icons.Rounded.Info,
-                                contentDescription = "About"
-                            )
-                        },
-                        onClick = {
-                            isDropdownVisible = false
-                            navController.navigate(AboutSettings)
-                        }
-                    )
                     DropdownMenuItem(
                         text = { Text("Experimental features") },
                         leadingIcon = {
@@ -281,328 +100,72 @@ fun PreferencesScreen(navController: NavHostController) {
                 .fillMaxSize()
                 .padding(it)
                 .nestedScroll(scrollBehavior.nestedScrollConnection),
-            contentPadding = PaddingValues(vertical = MEDIUM_SPACER.dp),
-            verticalArrangement = Arrangement.spacedBy(LARGE_SPACER.dp)
+            contentPadding = PaddingValues(MEDIUM_SPACER.dp)
         ) {
-            item {
-                ExpressiveGroup("General") {
-                    item {
-                        EnumPref(
-                            title = "Image source",
-                            summary = currentSettings.imageSource.label,
-                            enumItems = ImageSource.entries,
-                            selectedItem = currentSettings.imageSource,
-                            onSelection = {
-                                if (it == ImageSource.R34 && !AgeVerification.hasVerifiedAge(currentSettings)) {
-                                    showAgeVerificationDialog = true
-                                    return@EnumPref
-                                }
-                                scope.launch {
-                                    preferencesRepository.updatePref(
-                                        PreferenceKeys.IMAGE_SOURCE,
-                                        it
-                                    )
-                                }
-                                viewModel.clearTagSuggestions()
-                            }
-                        )
-                    }
-                    item {
-                        val authType = currentSettings.imageSource.imageBoard.apiKeyRequirement
-                        TitleSummary(
-                            modifier = Modifier
-                                .fillMaxWidth()
-                                .animateContentSize(),
-                            title = "Set API key",
-                            summary = if (authType != ImageBoardRequirement.NOT_NEEDED) {
-                                "${currentSettings.imageSource.label} requires an API key${if (authType == ImageBoardRequirement.RECOMMENDED) " for the best experience." else "."} " +
-                                "Tap to set."
-                            } else {
-                                "${currentSettings.imageSource.label} does not require an API key."
-                            },
-                            enabled = authType != ImageBoardRequirement.NOT_NEEDED
-                        ) {
-                            showAuthDialog = true
-                        }
-                    }
-                    item {
-                        SwitchPref(
-                            checked = currentSettings.saveSearchHistory,
-                            title = "Save search history",
-                            summary = "Save your 10 most recent searches. When this is disabled, your " +
-                                    "search history will be cleared and Breadboard will not save future " +
-                                    "searches."
-                        ) {
-                            scope.launch {
-                                if (!it) preferencesRepository.clearSearchHistory()
-                                preferencesRepository.updatePref(
-                                    key = PreferenceKeys.SAVE_SEARCH_HISTORY,
-                                    to = it
-                                )
-                            }
-                        }
-                    }
-                    item {
-                        SwitchPref(
-                            checked = currentSettings.useFixedLinks,
-                            title = "Share embeddable links",
-                            summary = "When sharing an image, use a link that embeds better on external platforms when possible.",
-                            infoText = "When embeddable links are enabled, sharing an image may " +
-                                       "use an alternative link depending on the source.\n\n" +
-                                       "Bilibili links are transformed into vxbilibili.com\n" +
-                                       "Bluesky links are transformed into fxbsky.app\n" +
-                                       "Pixiv links are transformed into phixiv.net\n" +
-                                       "Twitter links are transformed into fxtwitter.com\n" +
-                                       "Weibo links are transformed into fxweibo.com."
-                        ) {
-                            scope.launch {
-                                preferencesRepository.updatePref(
-                                    PreferenceKeys.USE_FIXED_LINKS,
-                                    it
-                                )
-                            }
-                        }
+            LazyExpressiveGroup(desiredTopPadding = null) {
+                item {
+                    TitleSummary(
+                        modifier = Modifier.fillMaxWidth(),
+                        title = "General",
+                        summary = "Image source, API keys, search history, and links.",
+                        leadingIcon = {
+                            ContainedIcon(imageVector = Icons.Rounded.Settings, label = "General")
+                        },
+                        trailingIcon = { ChevronRight() }
+                    ) {
+                        navController.navigate(GeneralSettings)
                     }
                 }
-            }
-
-            item {
-                ExpressiveGroup(title = "Content") {
-                    item {
-                        TitleSummary(
-                            modifier = Modifier.fillMaxWidth(),
-                            title = "Manage blocked tags",
-                            summary = "Add or remove tags to block from search results and recommendations.",
-                            trailingIcon = { ChevronRight() }
-                        ) {
-                            navController.navigate(BlockedTagsSettings)
-                        }
-                    }
-                    item {
-                        TitleSummary(
-                            modifier = Modifier.fillMaxWidth(),
-                            title = "Manage recommendations",
-                            summary = "Fine-tune your recommendations by customising which tags can be used to recommend new content.",
-                            trailingIcon = { ChevronRight() }
-                        ) {
-                            navController.navigate(RecommendationsSettings)
-                        }
-                    }
-                    item {
-                        SwitchPref(
-                            checked = currentSettings.excludeAi,
-                            title = "Hide AI-generated images",
-                            summary = "Attempt to hide AI-generated images by automatically adding " +
-                                    "AI-related tags to your block list."
-                        ) {
-                            scope.launch {
-                                preferencesRepository.updatePref(
-                                    PreferenceKeys.EXCLUDE_AI,
-                                    it
-                                )
-                            }
-                            viewModel.resetProviders()
-                        }
-                    }
-                    item {
-                        SwitchPref(
-                            checked = currentSettings.filterRatingsLocally,
-                            title = "Filter ratings locally",
-                            summary = "Rather than appending the selected ratings to the search query, " +
-                                      "filter the results by rating after searching.",
-                            infoText = "Danbooru limits searches to 2 tags " +
-                                       "(which includes ratings) without an API key.\n\n" +
-                                       "Enabling this option will allow you to filter by rating " +
-                                       "on all sources without an API key, and also let you filter " +
-                                       "ratings mid-search, but may cause less results to be shown " +
-                                       "at once, resulting in slightly higher data usage.\n\n" +
-                                       "Yande.re always requires this option.\n\n" +
-                                       "If you're unsure, keep this option enabled."
-                        ) {
-                            scope.launch {
-                                preferencesRepository.updatePref(
-                                    PreferenceKeys.FILTER_RATINGS_LOCALLY,
-                                    it
-                                )
-                            }
-                        }
-                    }
-                    item {
-                        EnumPref(
-                            title = "Automatically play videos",
-                            summary = currentSettings.autoplayVideos.label,
-                            infoText = "Control whether videos will automatically play when " +
-                                       "opening them.\n\n" +
-                                       "When this is enabled, videos will start muted until you " +
-                                       "manually unmute by tapping the icon or your device's " +
-                                       "volume up button.",
-                            enumItems = AutoplayVideosMode.entries,
-                            selectedItem = currentSettings.autoplayVideos,
-                            onSelection = {
-                                scope.launch {
-                                    preferencesRepository.updatePref(
-                                        PreferenceKeys.AUTOPLAY_VIDEOS,
-                                        it
-                                    )
-                                }
-                            }
-                        )
+                item {
+                    TitleSummary(
+                        modifier = Modifier.fillMaxWidth(),
+                        title = "Content",
+                        summary = "Blocked tags, recommendations, and video settings.",
+                        leadingIcon = {
+                            ContainedIcon(imageVector = Icons.Rounded.Visibility, label = "Content")
+                        },
+                        trailingIcon = { ChevronRight() }
+                    ) {
+                        navController.navigate(ContentSettings)
                     }
                 }
-            }
-
-            item {
-                ExpressiveGroup(title = "Interaction and layout") {
-                    item {
-                        EnumPref(
-                            title = "Start page",
-                            summary = currentSettings.defaultStartDestination.label,
-                            enumItems = StartDestination.entries,
-                            selectedItem = currentSettings.defaultStartDestination,
-                        ) {
-                            scope.launch {
-                                preferencesRepository.updatePref(
-                                    PreferenceKeys.DEFAULT_START_DESTINATION,
-                                    it
-                                )
-                            }
-                        }
-                    }
-                    /* It's an enum so EnumPref would be the "correct" implementation,
-                       but I think this is a better UX as there are currently only two options. */
-                    item {
-                        SwitchPref(
-                            checked = currentSettings.defaultBrowseTab == BrowseTab.FOLLOWING,
-                            title = "Prefer Following tab",
-                            summary = "By default, open the Following tab instead of the For You " +
-                                      "tab on the Browse page.",
-                        ) { following ->
-                            scope.launch {
-                                preferencesRepository.updatePref(
-                                    PreferenceKeys.DEFAULT_BROWSE_TAB,
-                                    if (following) BrowseTab.FOLLOWING else BrowseTab.FOR_YOU
-                                )
-                            }
-                        }
-                    }
-                    item {
-                        EnumPref(
-                            title = "Hide app content",
-                            summary = currentSettings.flagSecureMode.label,
-                            enumItems = FlagSecureMode.entries,
-                            selectedItem = currentSettings.flagSecureMode
-                        ) {
-                            scope.launch {
-                                preferencesRepository.updatePref(
-                                    PreferenceKeys.FLAG_SECURE_MODE,
-                                    it
-                                )
-                            }
-                        }
-                    }
-                    item {
-                        SwitchPref(
-                            checked = currentSettings.useStaggeredGrid,
-                            title = "Staggered grid",
-                            summary = "Use a staggered grid for images rather than a uniform grid."
-                        ) {
-                            scope.launch {
-                                preferencesRepository.updatePref(
-                                    PreferenceKeys.USE_STAGGERED_GRID,
-                                    it
-                                )
-                            }
-                        }
-                    }
-                    item {
-                        SwitchPref(
-                            checked = currentSettings.unifiedInfoSheet,
-                            title = "Classic info sheet",
-                            summary = "Display all art information in a single scrollable " +
-                                      "list rather than organising it into tabs."
-                        ) {
-                            scope.launch {
-                                preferencesRepository.updatePref(
-                                    PreferenceKeys.UNIFIED_INFO_SHEET,
-                                    it
-                                )
-                            }
-                        }
-                    }
-                    item {
-                        ReorderablePref(
-                            title = "Reorder image actions",
-                            dialogTitle = "Image actions",
-                            summary = "Customise the order of actions in the image viewer. The top " +
-                                    "action will be displayed separately in its own dedicated button.",
-                            items = currentSettings.imageViewerActions
-                        ) {
-                            scope.launch {
-                                preferencesRepository.updateEnumList(
-                                    PreferenceKeys.IMAGE_VIEWER_ACTION_ORDER,
-                                    it
-                                )
-                            }
-                        }
+                item {
+                    TitleSummary(
+                        modifier = Modifier.fillMaxWidth(),
+                        title = "Behaviour and layout",
+                        summary = "Start page, app layout, and image viewer actions.",
+                        leadingIcon = {
+                            ContainedIcon(imageVector = Icons.Rounded.Dashboard, label = "Behaviour and layout")
+                        },
+                        trailingIcon = { ChevronRight() }
+                    ) {
+                        navController.navigate(LayoutSettings)
                     }
                 }
-            }
-
-            item {
-                ExpressiveGroup("Data and storage") {
-                    item {
-                        EnumPref(
-                            title = "Data saver",
-                            summary = currentSettings.dataSaver.label,
-                            enumItems = DataSaver.entries,
-                            infoText = "When data saver is enabled, images will load in a " +
-                                       "lower resolution by default.\n\n" +
-                                       "Downloads will always be in the maximum resolution " +
-                                       "regardless of this setting.",
-                            selectedItem = currentSettings.dataSaver,
-                            onSelection = {
-                                scope.launch {
-                                    preferencesRepository.updatePref(
-                                        PreferenceKeys.DATA_SAVER,
-                                        it
-                                    )
-                                }
-                            }
-                        )
-                    }
-                    item {
-                        TitleSummary(
-                            modifier = Modifier.fillMaxWidth(),
-                            title = "Save downloads to",
-                            summary = if (currentSettings.storageLocation == Uri.EMPTY) "Tap to set"
-                            else currentSettings.storageLocation.toString()
-                        ) {
-                            storageLocationPromptLaunched = true
-                        }
+                item {
+                    TitleSummary(
+                        modifier = Modifier.fillMaxWidth(),
+                        title = "Data and storage",
+                        summary = "Download location and your Breadboard data.",
+                        leadingIcon = {
+                            ContainedIcon(imageVector = Icons.Rounded.Storage, label = "Data and storage")
+                        },
+                        trailingIcon = { ChevronRight() }
+                    ) {
+                        navController.navigate(DataSettings)
                     }
                 }
-            }
-
-            item {
-                ExpressiveGroup(title = "Import/export") {
-                    item {
-                        TitleSummary(
-                            modifier = Modifier.fillMaxWidth(),
-                            title = "Export data",
-                            summary = "Export a backup file containing your current settings, favourite images, " +
-                                    "and search history."
-                        ) {
-                            showExportDialog = true
-                        }
-                    }
-                    item {
-                        TitleSummary(
-                            modifier = Modifier.fillMaxWidth(),
-                            title = "Import data",
-                            summary = "Import a Breadboard backup file."
-                        ) {
-                            importingStarted = true
-                        }
+                item {
+                    TitleSummary(
+                        modifier = Modifier.fillMaxWidth(),
+                        title = "About",
+                        summary = "Breadboard info and licenses.",
+                        leadingIcon = {
+                            ContainedIcon(imageVector = Icons.Rounded.Info, label = "About")
+                        },
+                        trailingIcon = { ChevronRight() }
+                    ) {
+                        navController.navigate(AboutSettings)
                     }
                 }
             }
@@ -611,160 +174,48 @@ fun PreferencesScreen(navController: NavHostController) {
 }
 
 
+@OptIn(ExperimentalMaterial3ExpressiveApi::class)
 @Composable
-private fun AuthDialog(
-    selectedBoard: ImageBoard,
-    default: ImageBoardAuth?,
-    onDismissRequest: () -> Unit,
-    onSave: (String, String) -> Unit
+private fun ContainedIcon(
+    label: String,
+    imageVector: ImageVector
 ) {
-    var inQueryParamMode by remember { mutableStateOf(false) }
-    val queryRegex = remember { Regex("&api_key=([^&]+)&user_id=(\\d+)") }
-    var queryString by remember { mutableStateOf("") }
-    var isQueryStringValid by remember { mutableStateOf(false) }
+    val isDark = isSystemInDarkTheme()
+    // Assign the icon a random colour hue based on its label.
+    val seed = label.hashCode()
+    val hue = (seed and Integer.MAX_VALUE).rem(360).toFloat()
 
-    var userId by remember { mutableStateOf(default?.user ?: "") }
-    var apiKey by remember { mutableStateOf(default?.apiKey ?: "") }
+    /* We're going for light backgrounds with a darker inner icon.
+       The specific saturation and lightness depend on whether the user is in dark or light mode. */
+    val (bgSat, bgLight) = if (isDark) {
+        0.42f to 0.70f
+    } else {
+        0.65f to 0.87f
+    }
 
-    val context = LocalContext.current
+    val (iconSat, iconLight) = if (isDark) {
+        0.60f to 0.20f
+    } else {
+        0.65f to 0.27f
+    }
 
-    AlertDialog(
-        onDismissRequest = onDismissRequest,
-        title = {
-            Row(
-                modifier = Modifier.fillMaxWidth(),
-                horizontalArrangement = Arrangement.SpaceBetween,
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                Text("Set API key")
-                Box {
-                    var showMenu by remember { mutableStateOf(false) }
-                    IconButton(onClick = { showMenu = true }) {
-                        Icon(
-                            imageVector = Icons.Rounded.MoreVert,
-                            contentDescription = "Options"
-                        )
-                    }
-                    DropdownMenu(
-                        expanded = showMenu,
-                        onDismissRequest = { showMenu = false },
-                        containerColor = MaterialTheme.colorScheme.surfaceContainerHigh,
-                        shape = MaterialTheme.shapes.small
-                    ) {
-                        DropdownMenuItem(
-                            text = { Text(if (inQueryParamMode) "Enter credentials normally" else "Enter query string") },
-                            onClick = {
-                                inQueryParamMode = !inQueryParamMode
-                                showMenu = false
-                            }
-                        )
-                    }
-                }
-            }
-        },
-        text = {
-            Column {
-                if (inQueryParamMode) {
-                    PreferenceTextBox(
-                        value = queryString,
-                        label = "Query string",
-                        obscured = false,
-                        keyboardType = KeyboardType.Uri,
-                        isError = !isQueryStringValid && queryString.isNotEmpty()
-                    ) { newValue ->
-                        queryString = newValue.trim()
-                        isQueryStringValid = queryRegex.containsMatchIn(queryString)
-                    }
-                } else {
-                    PreferenceTextBox(
-                        value = userId,
-                        label = "User ID/name",
-                        obscured = false
-                    ) {
-                        userId = it.trim()
-                    }
-                }
+    val bgInt = ColorUtils.HSLToColor(floatArrayOf(hue, bgSat, bgLight))
+    val iconInt = ColorUtils.HSLToColor(floatArrayOf(hue, iconSat, iconLight))
 
-                AnimatedVisibility(!inQueryParamMode) {
-                    PreferenceTextBox(
-                        modifier = Modifier.padding(top = SMALL_SPACER.dp),
-                        value = apiKey,
-                        label = "API key",
-                        keyboardType = KeyboardType.Password,
-                        obscured = true
-                    ) {
-                        apiKey = it.trim()
-                    }
-                }
+    val bgColour = Color(bgInt)
+    val iconTint = Color(iconInt)
 
-                selectedBoard.apiKeyCreationUrl?.let { url ->
-                    val apiKeyCreationText = buildAnnotatedString {
-                        val link = LinkAnnotation.Url(
-                            url,
-                            TextLinkStyles(
-                                SpanStyle(color = MaterialTheme.colorScheme.secondary, textDecoration = TextDecoration.Underline)
-                            )
-                        ) {
-                            launchInWebBrowser(context, url)
-                        }
-
-                        withLink(link) {
-                            append("Find your credentials...")
-                        }
-                    }
-
-                    Text(
-                        text = apiKeyCreationText,
-                        style = MaterialTheme.typography.bodySmall,
-                        modifier = Modifier.padding(
-                            start = TINY_SPACER.dp,
-                            top = SMALL_SPACER.dp
-                        )
-                    )
-                }
-            }
-        },
-        confirmButton = {
-            Button(
-                enabled = if (inQueryParamMode) {
-                    isQueryStringValid || queryString.isBlank()
-                } else {
-                    (userId.isNotBlank() && apiKey.isNotBlank()) || (userId.isBlank() && apiKey.isBlank())
-                },
-                onClick = {
-                    if (inQueryParamMode) {
-                        if (queryString.isBlank()) {
-                            onSave("", "")
-                            return@Button
-                        }
-
-                        val match = queryRegex.find(queryString)
-                        if (match != null) {
-                            onSave(match.groupValues[2], match.groupValues[1])
-                        }
-
-                    } else {
-                        onSave(userId, apiKey)
-                    }
-                }
-            ) {
-                Text("Save")
-            }
-        },
-        dismissButton = {
-            TextButton(onClick = onDismissRequest) {
-                Text("Cancel")
-            }
-        }
-    )
-}
-
-
-private operator fun JSONObject.contains(key: String): Boolean {
-    return try {
-        get(key)
-        true
-    } catch (_: JSONException) {
-        false
+    Box(
+        modifier = Modifier
+            .clip(CircleShape)
+            .background(bgColour)
+            .fillMaxSize(),
+        contentAlignment = Alignment.Center
+    ) {
+        Icon(
+            imageVector = imageVector,
+            contentDescription = null,
+            tint = iconTint
+        )
     }
 }
