+++ app/src/main/java/moe/apex/breadboard/ui/theme/Type.kt
@@ -35,7 +35,7 @@ val Typography = Typography(
         titleMedium = Typography().titleMedium.copy(
                 fontFamily = FontFamily(Font(R.font.kumbh)),
                 fontWeight = FontWeight.Bold,
-                fontSize = 15.sp,
+                fontSize = 14.sp,
         ),
         headlineMedium = Typography().headlineMedium.copy(
                 fontFamily = FontFamily(Font(R.font.special_gothic_one_expanded)),
@@ -57,8 +57,8 @@ val Typography.searchField: TextStyle
 
 val Typography.prefTitle: TextStyle
         get() = Typography().bodyLarge.copy(
-                        fontSize = 17.sp,
-                        lineHeight = 17.sp,
+                        fontSize = 16.sp,
+                        lineHeight = 16.sp,
                         letterSpacing = 0.sp,
                         fontFamily = FontFamily(Font(R.font.kumbh)),
                         fontWeight = FontWeight.Bold
